backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.115.13
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data validation and serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# Background tasks
celery==5.3.4
redis==5.0.1

# HTTP client
httpx==0.25.2

# Environment variables
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# PDF processing (will be added later in PDF service)
# pdfplumber==0.10.3
# pytesseract==0.3.10
# opencv-python==4.8.1.78

# AI/ML (will be added later)
# google-generativeai==0.3.2